requirements.txt
albumentations==2.0.8
graphviz==0.20.3
matplotlib==3.7.1
matplotlib==3.8.2
MedPy==0.4.0
ml_collections==0.1.1
nibabel==5.3.2
numpy==1.21.6
numpy==1.24.3
opencv_python==4.5.5.64
opencv_python_headless==4.9.0.80
Pillow==10.0.1
Pillow==9.5.0
Pillow==11.3.0
scikit_image==0.19.3
scipy==1.8.1
SimpleITK==2.1.1.2
SimpleITK==2.5.2
tensorboardX==2.6.2.2
tensorboardX==2.6.4
torch==2.1.0
torch==2.0.0+cu118
torchviz==0.0.3
tqdm==4.63.2
tqdm==4.65.0
